+++ Sources/PatchDay/Sites/SitesTable.swift
@@ -28,21 +28,18 @@ class SitesTable: TableViewWrapper<SiteCell>, SitesTableProtocol {
             table.isEditing = false
             let endSiteIndex = (sites?.count ?? 1) - 1
             let endCellIndex = cellCount - 1
-            
+
             // Handle case where user deletes cell and then resets in same session
             if endCellIndex < endSiteIndex {
                 let insertRange = (endCellIndex + 1)...endSiteIndex
                 let indexPaths = insertRange.map({ (i: Index) -> IndexPath in IndexPath(row: i, section: 0) })
                 table.insertRows(at: indexPaths, with: .none)
             }
-            
+
             let range = 0...endSiteIndex
             let indexPathsToReload = range.map({ (i: Index) -> IndexPath in IndexPath(row: i, section: 0) })
             table.reloadRows(at: indexPathsToReload, with: .automatic)
-        }) {
-            _ in
-            self.correctCellProperties(startIndex: 0)
-        }
+        }, completion: {_ in self.correctCellProperties(startIndex: 0)})
     }
 
     subscript(index: Index) -> SiteCellProtocol {
@@ -68,7 +65,7 @@ class SitesTable: TableViewWrapper<SiteCell>, SitesTableProtocol {
             self.reloadCells()
         }
     }
-    
+
     private func createCellProps(_ siteIndex: Index) -> SiteCellProperties {
         var props = SiteCellProperties(row: siteIndex)
         guard let sites = sites else { return props }